+++ src/lib/Scenes/Fair/Components/SearchLink.tsx
@@ -14,10 +14,6 @@ export class SearchLink extends React.Component<Props> {
     SwitchBoard.presentNavigationViewController(this, `/${fairID}/search`)
   }
 
-  componentDidMount() {
-    console.log("mounted")
-  }
-
   render() {
     return (
       <TouchableWithoutFeedback onPress={this.handlePress}>