+++ CHANGELOG.md
@@ -34,6 +34,12 @@
 - Adds some small styling adjustments to the hoursCollapsible - kieran
 - Fix overfetching data on Fair Detail - luc
 - Update Show detail view spacing & copy - luc
+- Improves native integration with bottom sheet UI - ash
+- Adds missing fair booth separator - kieran
+- Refactors Fair Header view to use EntityList - luc
+- Fix bug with Markdown renderer - luc
+- Refactors Show Header view to use EntityList - luc
+- Fix fair booth view by adding query renderer - luc
 
 ### 1.8.1